Implement omit and pick operations for class-transformer.
While the `class-transformer` library implements a `MetadataStorage` abstraction,
it keeps this abstaction private, meaning that narrowing a transformer-enabled class
requires some other mechanism than metadata modification. Fortunately, there is
an `@Exclude()` decorator that we can programmatically apply to derived DTOs.
