+++ src/operators/__tests__/operators.test.ts
@@ -1,10 +1,12 @@
+import { plainToClass } from 'class-transformer';
+
 import { IsString } from '../../flavors/openapi';
 import {
     createOpenAPIDocument,
     createOpenAPIFixtureModule,
     parseOpenAPISchemas,
 } from '../../testing';
-import { derive, omit, pick } from '../class.operators';
+import { derive, omit, pick } from '..';
 
 class Foo {
     @IsString()
@@ -21,8 +23,9 @@ describe('class.operators', () => {
         });
     });
     describe('omit', () => {
+        const Bar = omit(Foo, 'Bar', ['baz']);
+
         it('picks a subset of OpenAPI properties', async () => {
-            const Bar = omit(Foo, 'Bar', ['baz']);
             const FixtureModule = createOpenAPIFixtureModule(Bar);
             const document = await createOpenAPIDocument(FixtureModule);
             const schemas = parseOpenAPISchemas(document);
@@ -37,10 +40,24 @@ describe('class.operators', () => {
                 },
             });
         });
+        it('picks a subset of fields to transform', () => {
+            const plain = {
+                bar: 'bar',
+                baz: 'baz',
+            };
+            expect(plainToClass(Foo, plain)).toEqual({
+                bar: 'bar',
+                baz: 'baz',
+            });
+            expect(plainToClass(Bar, plain)).toEqual({
+                bar: 'bar',
+            });
+        });
     });
     describe('pick', () => {
+        const Bar = pick(Foo, 'Bar', ['baz']);
+
         it('picks a subset of OpenAPI properties', async () => {
-            const Bar = pick(Foo, 'Bar', ['baz']);
             const FixtureModule = createOpenAPIFixtureModule(Bar);
             const document = await createOpenAPIDocument(FixtureModule);
             const schemas = parseOpenAPISchemas(document);
@@ -53,6 +70,19 @@ describe('class.operators', () => {
                         type: 'string',
                     }],
                 },
+            });
+        });
+        it('picks a subset of fields to transform', () => {
+            const plain = {
+                bar: 'bar',
+                baz: 'baz',
+            };
+            expect(plainToClass(Foo, plain)).toEqual({
+                bar: 'bar',
+                baz: 'baz',
+            });
+            expect(plainToClass(Bar, plain)).toEqual({
+                baz: 'baz',
             });
         });
     });
