+++ src/operators/index.ts
@@ -1,1 +1,3 @@
-export * from './class.operators';
+export * from './derive.operator';
+export * from './omit.operator';
+export * from './pick.operator';
